requirements.txt
# AlphaAgents Paper-Compliant Requirements
# AutoGen Studio (Paper's exact UI)
autogenstudio>=0.5.0

# LLM API
openai>=1.0.0

# Financial data
yfinance>=0.2.0
requests>=2.31.0
newsapi-python>=0.2.6

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Environment configuration
python-dotenv>=1.0.0